fix: move useJoinedColoniesWithExtension to useNetworkMotionStates as it close related functions
